requirements.txt
Django>=1.4
django-admin-bootstrapped>=0.3.2
django-modeltranslation>=0.6.1
